voice/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
SpeechRecognition==3.14.3
pydub==0.25.1
gTTS==2.4.0
pyttsx3==2.90
pyaudio==0.2.14
werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0